+++ paddleslim/dygraph/rep/reper/diversebranchblock.py
@@ -20,34 +20,10 @@ import paddle
 import paddle.nn as nn
 import paddle.nn.functional as F
 from paddle import ParamAttr
-from paddle.regularizer import L2Decay
-from paddle.nn.initializer import KaimingNormal
+
+from .base import BaseConv2DReper, ConvBNLayer
 
 __all__ = ["DiverseBranchBlock"]
-
-
-class ConvBNLayer(nn.Layer):
-    def __init__(self, in_channels, out_channels, filter_size, stride,
-                 groups=1):
-        super().__init__()
-        self.conv = nn.Conv2D(
-            in_channels=in_channels,
-            out_channels=out_channels,
-            kernel_size=filter_size,
-            stride=stride,
-            padding=(filter_size - 1) // 2,
-            groups=groups,
-            weight_attr=ParamAttr(initializer=KaimingNormal()),
-            bias_attr=False)
-        self.bn = nn.BatchNorm2D(
-            out_channels,
-            weight_attr=ParamAttr(regularizer=L2Decay(0.0)),
-            bias_attr=ParamAttr(regularizer=L2Decay(0.0)))
-
-    def forward(self, x):
-        x = self.conv(x)
-        x = self.bn(x)
-        return x
 
 
 class IdentityBasedConv1x1(nn.Conv2D):
@@ -124,7 +100,7 @@ class BNAndPadLayer(nn.Layer):
         return self.bn._epsilon
 
 
-class DiverseBranchBlock(nn.Layer):
+class DiverseBranchBlock(BaseConv2DReper):
     """
     An instance of the DBB module, which replaces the conv-bn layer in the network.
     """
@@ -135,14 +111,15 @@ class DiverseBranchBlock(nn.Layer):
                  kernel_size,
                  stride=1,
                  groups=1,
+                 padding=None,
                  internal_channels_1x1_3x3=None):
-        super(DiverseBranchBlock, self).__init__()
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.kernel_size = kernel_size
-        self.stride = stride
-        self.groups = groups
-        self.padding = (kernel_size - 1) // 2
+        super(DiverseBranchBlock, self).__init__(
+            in_channels=in_channels,
+            out_channels=out_channels,
+            kernel_size=kernel_size,
+            stride=stride,
+            groups=groups,
+            padding=padding)
 
         # kxk branch
         self.dbb_origin = ConvBNLayer(
